refactor: merging binary and library
Removing development of the library as it makes the release and development
process more complicated. Additionally, there are no users of the library and
all functionality should be covered by the binary.
