+++ Cargo.toml
@@ -1,15 +1,31 @@
-[workspace]
-resolver = "2"
-members = [
-    "is_affected",
-    "is_affected_lib",
-]
+[package]
+name = "is_affected"
+version = "0.4.2"
+description = "A utility for checking and listing the affected resources across a range of commits, useful when working with monorepos."
+authors = ["C <DeveloperC@protonmail.com>"]
+edition = "2021"
+license = "AGPL-3.0"
+repository = "https://github.com/DeveloperC286/is_affected/"
+readme = "README.md"
+keywords = ["monorepo", "commits", "affected", "repository", "git"]
+categories = ["command-line-utilities", "development-tools", "config", "parser-implementations", "text-processing"]
 
 
-[patch.crates-io]
-is_affected_lib = { path = "is_affected_lib" }
+[dependencies]
+# For logging.
+log = "0.4.20"
+pretty_env_logger = "0.5.0"
+
+# For CLI parsing.
+clap = { version = "4.4.6", features = ["derive"] }
+
+# For reading Git commits.
+git2 = { version = "0.17.2", default-features = false, features=[] }
+
+# For checking affected paths.
+regex = "1.9.6"
 
 
-[profile.release]
-lto = 'fat'
-codegen-units = 1
+[dev-dependencies]
+# For parameterized testing.
+rstest = "0.18.2"
