+++ ci/end-to-end-test.sh
@@ -3,5 +3,5 @@
 set -o errexit
 set -o xtrace
 
-cd "is_affected/end-to-end-tests/"
+cd "end-to-end-tests/"
 behave
