+++ ci/fix-python-formatting.sh
@@ -3,4 +3,4 @@
 set -o errexit
 set -o xtrace
 
-find "is_affected/end-to-end-tests/features/" -type f | grep "[.]py$" | xargs -I {} autopep8 --in-place --aggressive --aggressive --max-line-length 120 "{}"
+find "end-to-end-tests/features/" -type f | grep "[.]py$" | xargs -I {} autopep8 --in-place --aggressive --aggressive --max-line-length 120 "{}"
