+++ end-to-end-tests/features/from_shortened_commit_hash.feature
