+++ end-to-end-tests/features/is_affected_detects_directories.feature
