+++ end-to-end-tests/features/list_affected_resources.feature
