+++ end-to-end-tests/features/steps/then.py
@@ -39,7 +39,7 @@ def assert_is_not_affected(context):
 @then('their is a could not find commit hash "{commit_hash}" error.')
 def assert_could_not_find_commit_hash_error(context, commit_hash):
     # Given
-    could_not_find_commit_hash_error = f" ERROR is_affected_lib::commits > Can not find a commit with the hash '{commit_hash}'.\n"  # fmt: off
+    could_not_find_commit_hash_error = f" ERROR is_affected::commits > Can not find a commit with the hash '{commit_hash}'.\n"  # fmt: off
 
     # When/Then
     result = assert_is_not_affected(context)
@@ -53,7 +53,7 @@ def assert_could_not_find_commit_hash_error(context, commit_hash):
 def assert_could_not_find_shortened_commit_hash_error(
         context, shortened_commit_hash):
     # Given
-    could_not_find_shortened_commit_hash_error = f" ERROR is_affected_lib::commits > No commit hashes start with the provided short commit hash \"{shortened_commit_hash}\".\n"  # fmt: off
+    could_not_find_shortened_commit_hash_error = f" ERROR is_affected::commits > No commit hashes start with the provided short commit hash \"{shortened_commit_hash}\".\n"  # fmt: off
 
     # When/Then
     result = assert_is_not_affected(context)
@@ -67,7 +67,7 @@ def assert_could_not_find_shortened_commit_hash_error(
 def assert_ambiguous_shortened_commit_hash_error(
         context, shortened_commit_hash):
     # Given
-    ambiguous_shortened_commit_hash_error = re.compile(f"^ ERROR is_affected_lib::commits > Ambiguous short commit hash, the commit hashes [[]({shortened_commit_hash}[a-f0-9]*(, )?)*[]] all start with the provided short commit hash \"{shortened_commit_hash}\".\n$")  # fmt: off
+    ambiguous_shortened_commit_hash_error = re.compile(f"^ ERROR is_affected::commits > Ambiguous short commit hash, the commit hashes [[]({shortened_commit_hash}[a-f0-9]*(, )?)*[]] all start with the provided short commit hash \"{shortened_commit_hash}\".\n$")  # fmt: off
 
     # When/Then
     result = assert_is_not_affected(context)
@@ -79,7 +79,7 @@ def assert_ambiguous_shortened_commit_hash_error(
 @then('their is a could not find reference "{reference}" error.')
 def assert_could_not_find_reference_error(context, reference):
     # Given
-    could_not_find_reference_error = f" ERROR is_affected_lib::commits > Could not find a reference with the name \"{reference}\".\n"  # fmt: off
+    could_not_find_reference_error = f" ERROR is_affected::commits > Could not find a reference with the name \"{reference}\".\n"  # fmt: off
 
     # When/Then
     result = assert_is_not_affected(context)
