--- is_affected_lib/.gitignore
@@ -1,13 +1,0 @@
-# Rust .gitignore
-# Generated by Cargo
-debug/
-target/
-
-# Generated by rustfmt
-**/*.rs.bk
-
-# Generated by proptest
-proptest-regressions/
-
-# Generated by IntelliJ IDEA
-.idea/
