--- is_affected_lib/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "is_affected_lib"
-version = "0.1.2"
-description = "A utility for checking and listing the affected resources across a range of commits, useful when working with monorepos."
-authors = ["C <DeveloperC@protonmail.com>"]
-edition = "2021"
-license = "AGPL-3.0"
-repository = "https://github.com/DeveloperC286/is_affected/"
-readme = "README.md"
-keywords = ["monorepo", "commits", "affected", "repository", "git"]
-categories = ["data-structures", "development-tools", "config", "parser-implementations", "text-processing"]
-
-
-[dependencies]
-# For logging.
-log = "0.4.20"
-
-# For reading Git commits.
-git2 = { version = "0.17.2", default-features = false, features=[] }
-
-# For checking affected paths.
-regex = "1.9.6"
-
-
-[dev-dependencies]
-# For parameterized testing.
-rstest = "0.18.2"
