--- is_affected_lib/README.md
@@ -1,26 +1,0 @@
-> 🚧 Archived 🚧
-> Please use the binary version instead, it should contain all functionality you were going to implement with the library.
-> * https://github.com/DeveloperC286/is_affected
-
-[![crates.io](https://img.shields.io/crates/v/is_affected_lib)](https://crates.io/crates/is_affected_lib)
-[![Conventional Commits](https://img.shields.io/badge/Conventional%20Commits-1.0.0-yellow.svg)](https://conventionalcommits.org)
-[![License](https://img.shields.io/badge/License-AGPLv3-blue.svg)](https://www.gnu.org/licenses/agpl-3.0)
-
-
-A utility for checking and listing the affected resources across a range of commits, useful when working with monorepos.
-
-
-## Content
- * [Usage](#usage)
-   + [Usage - Logging](#usage-logging)
- * [Issues/Feature Requests](#issuesfeature-requests)
-
-
-## Usage
-### Usage - Logging
-The crate `log` is used to provide logging.
-See [https://crates.io/crates/log](https://crates.io/crates/log) for more detailed documentation.
-
-
-## Issues/Feature Requests
-To report an issue or request a new feature use [https://github.com/DeveloperC286/is_affected/issues](https://github.com/DeveloperC286/is_affected/issues).
